Split streamlit to classes.
